wip: restructured code to make it easier to test
