+++ lib/common.js
@@ -10,6 +10,7 @@ const common = module.exports
 common.PACKAGE_ROOT = path.resolve(__dirname, '..')
 common.BUILD_PATH = path.resolve(common.PACKAGE_ROOT, './build/Release')
 common.REMOTE_PATH = process.env.NR_NATIVE_METRICS_REMOTE_PATH || 'nodejs_agent/builds/'
+common.IS_WIN = process.platform === 'win32'
 
 common.parseArgs = function parseArgs(_argv, _opts) {
   const args = []
