+++ lib/pre-build.js
@@ -13,7 +13,6 @@
 // XXX This file must not have any deps. This file will run during the install
 // XXX step of the module and we are _not_ guaranteed that the dependencies have
 // XXX already installed. Core modules are okay.
-const cp = require('child_process')
 const fs = require('fs')
 const http = require('http')
 const https = require('https')
@@ -31,28 +30,23 @@ const {
   logFinish,
   PACKAGE_ROOT,
   BUILD_PATH,
-  REMOTE_PATH
+  REMOTE_PATH,
+  IS_WIN
 } = require('./common')
+const { execGyp, gypVersion } = require('./gyp-utils')
 
 const CPU_COUNT = os.cpus().length
-const IS_WIN = process.platform === 'win32'
 const DOWNLOAD_HOST =
   process.env.NR_NATIVE_METRICS_DOWNLOAD_HOST || 'https://download.newrelic.com/'
 
 const opts = {}
-exports.load = load
-exports.executeCli = executeCli
+const preBuild = module.exports
 
-if (require.main === module) {
-  const [, , cmd, target] = parseArgs(process.argv, opts)
-  executeCli(cmd, target)
-}
-
-function load(target) {
+preBuild.load = function load(target) {
   return require(path.join(BUILD_PATH, getBinFileName(target)))
 }
 
-function makePath(pathToMake, cb) {
+preBuild.makePath = function makePath(pathToMake, cb) {
   const accessRights = fs.constants.R_OK | fs.constants.W_OK
 
   // We only want to make the parts after the package directory.
@@ -88,99 +82,7 @@ function makePath(pathToMake, cb) {
   }
 }
 
-function findNodeGyp() {
-  // This code heavily borrows from node-pre-gyp.
-  // https://github.com/mapbox/node-pre-gyp/blob/e0b3b6/lib/util/compile.js#L18-L55
-
-  // First, look for it in the NPM environment variable.
-  let gypPath = null
-  if (process.env.npm_config_node_gyp) {
-    try {
-      gypPath = process.env.npm_config_node_gyp
-      fs.accessSync(gypPath)
-      return gypPath
-    } catch (err) {
-      // This method failed, hopefully the next will succeed...
-    }
-  }
-
-  // Next, see if the package is installed somewhere.
-  try {
-    // eslint-disable-next-line node/no-missing-require
-    const gypPkgPath = require.resolve('node-gyp')
-    gypPath = path.resolve(gypPkgPath, '../../bin/node-gyp.js')
-    fs.accessSync(gypPath)
-    return gypPath
-  } catch (err) {
-    // This method failed, hopefully the next will succeed...
-  }
-
-  // Then look for it in NPM's install location.
-  try {
-    // eslint-disable-next-line node/no-missing-require
-    const npmPkgPath = require.resolve('npm')
-    gypPath = path.resolve(npmPkgPath, '../../node_modules/node-gyp/bin/node-gyp.js')
-    fs.accessSync(gypPath)
-    return gypPath
-  } catch (err) {
-    // This method failed, hopefully the next will succeed...
-  }
-
-  // All of that failed, now look for it next to node itself.
-  const nodeNpmPkgPath = path.resolve(process.execPath, '../../lib/node_modules/npm/')
-  gypPath = path.join(nodeNpmPkgPath, 'node_modules/node-gyp/bin/node-gyp.js')
-  try {
-    fs.accessSync(gypPath)
-    return gypPath
-  } catch {
-    return null
-  }
-}
-
-function gypVersion() {
-  let cmd = null
-  const args = ['-v']
-  const gyp = findNodeGyp()
-  if (gyp) {
-    args.unshift(gyp) // push_front
-    cmd = process.execPath
-  } else {
-    cmd = IS_WIN ? 'node-gyp.cmd' : 'node-gyp'
-  }
-
-  const child = cp.spawnSync(cmd, args)
-  const match = /v(\d+\.\d+\.\d+)/.exec(child.stdout)
-  return match && match[1]
-}
-
-function execGyp(args, cb) {
-  let cmd = null
-  const gyp = findNodeGyp()
-  if (gyp) {
-    args.unshift(gyp) // push_front
-    cmd = process.execPath
-  } else {
-    cmd = IS_WIN ? 'node-gyp.cmd' : 'node-gyp'
-  }
-
-  const spawnOpts = {}
-  if (!opts.quiet) {
-    spawnOpts.stdio = [0, 1, 2]
-  }
-  console.log('> ' + cmd + ' ' + args.join(' ')) // eslint-disable-line no-console
-
-  const child = cp.spawn(cmd, args, spawnOpts)
-  child.on('error', cb)
-  child.on('close', function onGypClose(code) {
-    if (code !== 0) {
-      cb(new Error('Command exited with non-zero code: ' + code))
-    } else {
-      cb(null)
-    }
-  })
-}
-
-function build(target, rebuild, cb) {
+preBuild.build = function build(target, rebuild, cb) {
   const HAS_OLD_NODE_GYP_ARGS_FOR_WINDOWS = semver.lt(gypVersion() || '0.0.0', '3.7.0')
 
   if (IS_WIN && HAS_OLD_NODE_GYP_ARGS_FOR_WINDOWS) {
@@ -189,13 +91,13 @@ function build(target, rebuild, cb) {
 
   const cmds = rebuild ? ['clean', 'configure'] : ['configure']
 
-  execGyp(cmds, function cleanCb(err) {
+  execGyp(cmds, opts, function cleanCb(err) {
     if (err) {
       return cb(err)
     }
 
     const jobs = Math.round(CPU_COUNT / 2)
-    execGyp(['build', '-j', jobs, target], cb)
+    execGyp(['build', '-j', jobs, target], opts, cb)
   })
 }
 
@@ -205,7 +107,7 @@ function moveBuild(target, cb) {
   fs.rename(filePath, destination, cb)
 }
 
-function download(target, cb) {
+preBuild.download = function download(target, cb) {
   /* eslint-disable no-console */
   const fileName = getPackageFileName(target)
   const url = DOWNLOAD_HOST + REMOTE_PATH + fileName
@@ -272,8 +174,8 @@ function download(target, cb) {
   /* eslint-enable no-console */
 }
 
-function saveDownload(target, data, cb) {
-  makePath(BUILD_PATH, function makePathCB(err) {
+preBuild.saveDownload = function saveDownload(target, data, cb) {
+  preBuild.makePath(BUILD_PATH, function makePathCB(err) {
     if (err) {
       return cb(err)
     }
@@ -283,7 +185,7 @@ function saveDownload(target, data, cb) {
   })
 }
 
-function install(target, cb) {
+preBuild.install = function install(target, cb) {
   const errors = []
 
   const noBuild = opts['no-build'] || process.env.NR_NATIVE_METRICS_NO_BUILD
@@ -296,7 +198,7 @@ function install(target, cb) {
 
   // Otherwise, first attempt to build the package using the source. If that fails, try
   // downloading the package. If that also fails, whoops!
-  build(target, true, function buildCB(buildErr) {
+  preBuild.build(target, true, function buildCB(buildErr) {
     if (!buildErr) {
       return moveBuild(target, function moveBuildCB(moveErr) {
         if (moveErr) {
@@ -318,12 +220,12 @@ function install(target, cb) {
       return doCallback(new Error('Downloading is disabled.'))
     }
 
-    download(target, function downloadCB(err, data) {
+    preBuild.download(target, function downloadCB(err, data) {
       if (err) {
         return doCallback(err)
       }
 
-      saveDownload(target, data, doCallback)
+      preBuild.saveDownload(target, data, doCallback)
     })
   }
 
@@ -337,10 +239,10 @@ function install(target, cb) {
   }
 }
 
-function executeCli(cmd, target) {
+preBuild.executeCli = function executeCli(cmd, target) {
   logStart(cmd)
   if (cmd === 'build' || cmd === 'rebuild') {
-    build(target, cmd === 'rebuild', function buildCb(err) {
+    preBuild.build(target, cmd === 'rebuild', function buildCb(err) {
       if (err) {
         logFinish(cmd, target, err)
       } else {
@@ -348,6 +250,11 @@ function executeCli(cmd, target) {
       }
     })
   } else if (cmd === 'install') {
-    install(target, logFinish.bind(this, cmd, target))
+    preBuild.install(target, logFinish.bind(this, cmd, target))
   }
+}
+
+if (require.main === module) {
+  const [, , cmd, target] = parseArgs(process.argv, opts)
+  preBuild.executeCli(cmd, target)
 }
