+++ third_party_manifest.json
@@ -1,5 +1,5 @@
 {
-  "lastUpdated": "Wed Jun 21 2023 13:39:33 GMT-0400 (Eastern Daylight Time)",
+  "lastUpdated": "Wed Jun 28 2023 16:53:28 GMT-0400 (Eastern Daylight Time)",
   "projectName": "Native Metrics for New Relic Node Agent",
   "projectUrl": "https://github.com/newrelic/node-native-metrics",
   "includeOptDeps": false,
